backend/requirements.txt
<<<<<<< HEAD
youtube-transcript-api==1.2.4
pytube==15.0.0
=======
faiss-cpu>=1.8.0
google-genai>=0.7.0
langchain-text-splitters>=0.2.2
nltk>=3.9
numpy>=1.26.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.1
streamlit>=1.36.0
transformers>=4.42.0
youtube-transcript-api>=1.2.4
>>>>>>> 0443724f3fdec2f72b473196be5cc51b69a376bc
